+++ src/main/java/engima/waratsea/model/ship/AircraftCarrier.java
@@ -33,7 +33,6 @@ import lombok.Getter;
 import lombok.Setter;
 import org.jetbrains.annotations.NotNull;
 
-import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -321,7 +320,7 @@ public class AircraftCarrier implements Ship, Airbase {
      * @return The current number of steps deployed at this air base.
      */
     @Override
-    public BigDecimal getCurrentSteps() {
+    public int getCurrentSteps() {
         return squadrons.getCurrentSteps();
     }
 
@@ -544,7 +543,7 @@ public class AircraftCarrier implements Ship, Airbase {
     }
 
     /**
-     * Upddate the patrol.
+     * Update the patrol.
      *
      * @param patrolType The type of patrol.
      * @param patrolSquadrons  The squadrons on patrol.
@@ -664,7 +663,7 @@ public class AircraftCarrier implements Ship, Airbase {
      * @return True if this squadrons carrier has room for the given squadron. False otherwise.
      */
     private boolean hasRoom(final Squadron squadron) {
-        int steps = squadron.getSteps().intValue();
+        int steps = squadron.getSteps();
         return steps + deployedSteps() <= getMaxCapacity();
     }
