+++ src/main/java/engima/waratsea/presenter/squadron/Deployment.java
@@ -5,13 +5,11 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 
-import java.math.BigDecimal;
-
 @Slf4j
 public class Deployment {
     @Getter private final LandingType type;
-    @Getter @Setter private BigDecimal totalSteps;
-    @Getter @Setter private BigDecimal deployedSteps;
+    @Getter @Setter private int totalSteps;
+    @Getter @Setter private int deployedSteps;
 
     /**
      * The constructor.
